move welcome message to command/init
